# Conflicts:
#	WindowsFormsApp2/BUS/bin/Debug/BUS.dll
#	WindowsFormsApp2/BUS/bin/Debug/BUS.pdb
#	WindowsFormsApp2/BUS/bin/Debug/DAO.dll
#	WindowsFormsApp2/BUS/bin/Debug/DAO.pdb
#	WindowsFormsApp2/BUS/obj/Debug/BUS.csprojAssemblyReference.cache
#	WindowsFormsApp2/BUS/obj/Debug/BUS.dll
#	WindowsFormsApp2/BUS/obj/Debug/BUS.pdb
#	WindowsFormsApp2/DAO/bin/Debug/DAO.dll
#	WindowsFormsApp2/DAO/bin/Debug/DAO.pdb
#	WindowsFormsApp2/DAO/obj/Debug/DAO.csprojAssemblyReference.cache
#	WindowsFormsApp2/DAO/obj/Debug/DAO.dll
#	WindowsFormsApp2/DAO/obj/Debug/DAO.pdb
#	WindowsFormsApp2/WindowsFormsApp2/Main.Designer.cs
#	WindowsFormsApp2/WindowsFormsApp2/Main.cs
#	WindowsFormsApp2/WindowsFormsApp2/bin/Debug/BUS.dll
#	WindowsFormsApp2/WindowsFormsApp2/bin/Debug/BUS.pdb
#	WindowsFormsApp2/WindowsFormsApp2/bin/Debug/DAO.dll
#	WindowsFormsApp2/WindowsFormsApp2/bin/Debug/DAO.pdb
#	WindowsFormsApp2/WindowsFormsApp2/bin/Debug/WindowsFormsApp2.exe
#	WindowsFormsApp2/WindowsFormsApp2/bin/Debug/WindowsFormsApp2.pdb
#	WindowsFormsApp2/WindowsFormsApp2/obj/Debug/Presentation.csproj.GenerateResource.cache
#	WindowsFormsApp2/WindowsFormsApp2/obj/Debug/Presentation.csprojAssemblyReference.cache
#	WindowsFormsApp2/WindowsFormsApp2/obj/Debug/WindowsFormsApp2.exe
#	WindowsFormsApp2/WindowsFormsApp2/obj/Debug/WindowsFormsApp2.pdb

+++ WindowsFormsApp2/DAO/HangDAO.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -38,13 +39,15 @@ namespace DAO
             DataTable dt = this.dp.ExecuteQuery(query);
 
             List<HangDTO> hangS = new List<HangDTO>();
+            
             foreach (DataRow dr in dt.Rows)
             {
                 HangDTO hang = new HangDTO((int)dr["MAHANG"], (int)dr["MANVPHUTRACH"], (string)dr["TENHANG"],
-                    (int)dr["SOLUONGCONLAI"], (float)dr["DONGIA"]);
+                    (int)dr["SOLUONGCONLAI"], (double)dr["DONGIA"]);
                 
                 hangS.Add(hang);
-            }
+            }                
+                   
             return hangS;
         }
     }
